% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_regions.R
\name{define_regions}
\alias{define_regions}
\title{Run mutation region definition and annotation}
\usage{
define_regions(mut, reg_size = 100)
}
\arguments{
\item{mut}{a data.frame like object with mutations}
}
\description{
Run mutation region definition and annotation
}
