% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_panel.R
\name{eval_panel}
\alias{eval_panel}
\title{Evaluate and annotate a given set of regions (panel)}
\usage{
eval_panel(reg_to_patient_mut)
}
\arguments{
\item{reg_to_patient_mut}{A data.frame like object from \code{panel_to_patient()}
with all regions \code{reg_id} linked to patients \code{patient_id} via mutations \code{mut_id}.
It should additionally contain the following columns:
\itemize{
\item \code{reg_chr}
\item \code{reg_start} one-based
\item \code{reg_end}  one-based
}}
}
\value{
A data frame with a single row and annotation columns to the panel
}
\description{
Evaluate and annotate a given set of regions (panel)
}
\examples{

r2p <- panel_to_patient(gr_toy, mut_toy)
r2p

eval_panel(r2p)

}
