% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_region.R
\name{eval_regions}
\alias{eval_regions}
\title{Evaluate and annotate a given region to patient association data set}
\usage{
eval_regions(reg_to_patient_mut)
}
\arguments{
\item{reg_to_patient_mut}{A data.frame like object from \code{panel_to_patient()}
with all regions \code{reg_id} linked to patients \code{patient_id} via mutations \code{mut_id}.
It should additinoally countain the following columns:
\itemize{
\item \code{reg_chr}
\item \code{reg_start} one-based
\item \code{reg_end}  one-based
}}
}
\value{
A data frame with all unique regions and additional annotations.
}
\description{
Evaluate and annotate a given region to patient association data set
}
\examples{

r2p <- panel_to_patient(gr_toy, mut_toy)
r2p

eval_regions(r2p)

}
