% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_regions.R
\name{get_exons}
\alias{get_exons}
\title{Get all exons as GRanges object}
\usage{
get_exons(edb, columns = c("tx_id", "gene_id", "symbol", "gene_name"), ...)
}
\arguments{
\item{edb}{An Ensbel database object (EnsDb for Ensembl), e.g. from EnsDb.Hsapiens.v75}

\item{columns}{Annotation columns. See \code{ensembldb::exons()}}

\item{...}{Additional paramters passed to \code{ensembldb::exons()}}
}
\value{
A GenomicRanges::GRanges() object with regions.
}
\description{
Get all exons as GRanges object
}
\examples{
require(EnsDb.Hsapiens.v75)
edb <- EnsDb.Hsapiens.v75
get_exons(edb)

}
