% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr_toy}
\alias{gr_toy}
\title{A toy example genomic ranges object.}
\format{
A \code{\link[GenomicRanges]{GRanges}} object with rows 2 ranges.
}
\source{
made up
}
\usage{
gr_toy
}
\description{
This example \emph{panel} contains regions overlapping toy example mutations \code{m1}
and \code{m2} from \code{\link{mut_toy}}.
}
\keyword{datasets}
