% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_regions.R
\name{mut_to_gr}
\alias{mut_to_gr}
\title{Create GRanges object out of mutations}
\usage{
mut_to_gr(mut)
}
\arguments{
\item{mut}{A data.frame like object with all mutations and patients of the
cohort of interest. Should have the following columns:
\itemize{
\item \code{patient_id}
\item \code{mut_id}
\item \code{chr}
\item \code{start} or \code{pos}
\item \code{end} or \code{pos}
}}
}
\description{
Create GRanges object out of mutations
}
\examples{
data(mut_toy)
mut_to_gr(mut_toy)

}
