% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_regions.R
\name{mut_to_reg_gr}
\alias{mut_to_reg_gr}
\title{Create unique regions of fixed size around mutations as GRanges}
\usage{
mut_to_reg_gr(mut, reg_size = 100)
}
\arguments{
\item{mut}{A data.frame like object with all mutations and patients of the
cohort of interest. Should have the following columns:
\itemize{
\item \code{patient_id}
\item \code{mut_id}
\item \code{chr}
\item \code{start} or \code{pos}
\item \code{end} or \code{pos}
}}

\item{reg_size}{integer size of the region around mutations}
}
\description{
Create unique regions of fixed size around mutations as GRanges
}
\examples{
data(mut_toy)
mut_to_reg_gr(mut_toy)
mut_to_reg_gr(mut_toy, reg_size = 50)

}
