% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_to_patient.R
\name{panel_to_patient}
\alias{panel_to_patient}
\title{Get mapping of panel regions to covered patients and mutations.}
\usage{
panel_to_patient(reg_gr, mut)
}
\arguments{
\item{reg_gr}{A GRanges object with regions defining the panel. Should have
unique names and is expected to be sorted by decreasing priority.}

\item{mut}{A data.frame like object with all mutations and patients of the
cohort of interest. Should have the following columns:
\itemize{
\item \code{patient_id}
\item \code{mut_id}
\item \code{chr}
\item \code{start}
\item \code{end}
}}
}
\value{
A data set mapping the regions to affected patients and mutations.
The output contains at least the following columns
\itemize{
\item \code{reg_id}
\item \code{mut_id}
\item \code{patient_id}
}
}
\description{
Get mapping of panel regions to covered patients and mutations.
}
\examples{
gr_toy
mut_toy

panel_to_patient(gr_toy, mut_toy)

}
