% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_panel.R
\name{reduce_panel}
\alias{reduce_panel}
\title{Reduce panel regions to input target ranges by respecting a minimum region size.}
\usage{
reduce_panel(reg_gr, target_gr, min_size = 100)
}
\arguments{
\item{reg_gr}{A GRanges object with regions defining the panel.}

\item{target_gr}{A GRanges object with regions defining the target regions to
which the panel should be reduced.}

\item{min_size}{integer size in bp as the minimal size an individual region
can be reduced to.

In cases where the overlapping part of an input region with a target region
is smaller than \code{min_size}, the overlapping part will be resized to \code{min_size}
by fixing the center of the overlapping part.}
}
\description{
Reduce panel regions to input target ranges by respecting a minimum region size.
}
\examples{
reg_gr <- GenomicRanges::GRanges(c(
  "1:100-500",
  "1:700-900",
  "1:1100-1200"
))
target_gr <- GenomicRanges::GRanges(c(
  "1:200-600",
  "1:800-1000"
))

r <- reduce_panel(reg_gr, target_gr, min_size = 100)

}
