% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mut_toy}
\alias{mut_toy}
\title{A toy example data set linking patients to mutations.}
\format{
A data frame with rows and 5 colums:
\describe{
\item{patient_id}{patient ID}
\item{mut_id}{unique mutation ID}
\item{chr}{chromosome name}
\item{start}{start coordinate of the mutation}
\item{end}{end coordinate of the mutation}
}
}
\source{
made up
}
\usage{
mut_toy
}
\description{
A toy example data set linking patients to mutations.
}
\keyword{datasets}
