% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_greedy.R
\name{select_greedy}
\alias{select_greedy}
\title{Select mutaitons by maximal coverage}
\usage{
select_greedy(set_df, mut_col = "mut_id", sample_col = "patient_id")
}
\arguments{
\item{set_df}{A dataset of mutations (sets) in patients. Each line is an
occuracne of a mutation in a patient.}

\item{mut_col}{Character for column name of the mutation identifier (set)}

\item{sample_col}{Character for column name of the sample identifier (elements in set)}
}
\value{
a \link[tibble:tibble-package]{tibble} with the selected mutations (sets), number and coverge of the mutaitons
}
\description{
Select mutaitons by maximal coverage
}
